+++ dev-packages/private-eslint-plugin/index.js
@@ -21,6 +21,7 @@ exports.rules = {
     "localization-check": require('./rules/localization-check'),
     "named-logger-check": require('./rules/named-logger-check'),
     "no-src-import": require('./rules/no-src-import'),
+    "preload-localization-check": require('./rules/preload-localization-check'),
     "runtime-import-check": require('./rules/runtime-import-check'),
     "shared-dependencies": require('./rules/shared-dependencies')
 };